+++ src/components/form-elements/error-message/ErrorMessage.tsx
@@ -11,7 +11,7 @@ export const ErrorMessage: FC<ErrorMessageProps> = ({
   children,
   ...rest
 }) => {
-  if (!children || typeof children !== 'string') {
+  if (!children) {
     return null;
   }