+++ src/util/types/FormTypes.ts
@@ -1,3 +1,5 @@
+import { type ReactElement } from 'react';
+
 import { type ErrorMessageProps } from '#components/form-elements/error-message/index.js';
 import { type FieldsetProps } from '#components/form-elements/fieldset/index.js';
 import { type HintTextProps } from '#components/form-elements/hint-text/index.js';
@@ -7,13 +9,13 @@ import { type ComponentPropsWithDataAttributes } from '#util/types/NHSUKTypes.js
 
 export interface FormElementProps {
   'fieldsetProps'?: FieldsetProps;
-  'legend'?: string;
+  'legend'?: string | ReactElement;
   'legendProps'?: LegendProps;
-  'label'?: string;
+  'label'?: string | ReactElement;
   'labelProps'?: LabelProps;
-  'error'?: string;
+  'error'?: string | ReactElement;
   'errorProps'?: ErrorMessageProps;
-  'hint'?: string;
+  'hint'?: string | ReactElement;
   'hintProps'?: HintTextProps;
   'formGroupProps'?: ComponentPropsWithDataAttributes<'div'>;
   'id'?: string;